ENGR00175261 - EPDC fb: Remove warning messages from EPDC init
- Remove screen update from probe function.  This update is redundant.
The update in mxc_epdc_fb_fw_handler() will update and show the tux
logo if FB console is added, so this achieves the same effect as the
update invocation that was removed.
- Remove dmaengine_get(), because SDMA driver returns a failure when
cycling through DMA channels.  Since it is not essential to register
with this call, it has been removed.  dmaengine_put() also removed.
- Added hw_inialization variable to track whether HW is in process
of being initialized.  In which case, we do not print an error
message when an update is received.

Signed-off-by: Danny Nold <dannynold@freescale.com>