ENGR00283012 dma: imx-sdma: Drop unstable dma_free_coherent() in probe()
Because we declared the memory property when allocating it, after doing
dma_free_coherent(), the memory property might be not properly cleared.
If there're another driver allocate this part of area by asserting a
different property, there might be a kernel crash happen to this case.

As a work around, dropping this would cause tiny memory leak if probe()
of sdma driver failed, which should not happen regularly though.

Acked-by: Huang Shijie <b32955@freescale.com>
Signed-off-by: Nicolin Chen <b42378@freescale.com>