ENGR00309951 MX6sl EVK: improve PWM bl period based on tests
There is a SEIKO WVGA panel on the MX6sl EVK platform.
PWM1 is used to control the panel's backlight brightness by
connecting the VLED+ and VLED- pins of the panel after some
hardware conversion logic on board.  The panel's spec doesn't
mention if it's backlight brightness can be tuned or not.
Anyway, the brightness can be tuned via the PWM approach,
based on experiments.  An alternative is to add a new GPIO
backlight device driver for this.  But, a panel with a
configurable backlight brightness is necessary for a smart
device, for example, on the Android platform.  The original
PWM backlight period is too short for the panel, which causes
the brightness can not be steadily increased or decreased if
we change the brightness in one direction.  This patch
improves the period from 50000ns to 1000000ns.  The new period
is an value based on tests.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>