MLK-9685 ARM:imx6x:Ensure that the kernel can boot with "nosmp" in the command line
When nosmp is added to the command line, setup_max_cpus is set to 0
by the kernel. And this results in num_possible_cpus() returning 0 and the kernel
does not boot.
This patch ensures that at least one CPU's state is set to "possible" as part
of the boot process.
Cherry picked from 3.10.y kernel:
e94363b07a94bfe587a78330f7b9517f5cacb273

Signed-off-by: Ranjani Vaidyanathan <Ranjani.Vaidyanathan@freescale.com>